Add multiple reference genome folders support. Add a folder in input to store user provide genomes.

+++ src/check_contig.py
@@ -54,25 +54,19 @@ def _main():
 	#Use argparser.
 	parser = argparse.ArgumentParser( description = "Check genomes files and filtered out too short contigs. No short contig means creating symlink to the raw file. Filtered genome will be placed in the path users can specify." )
 	parser.add_argument( "-i", metavar = "input_ref", dest = "input_ref", required = True, help = "input converted abundance files" )
-	parser.add_argument( "-g", metavar = "genome_ref_path", dest = "rawGenome_path", required = True, help = "input path to IMG reference genome files" )
+	parser.add_argument( "-g", metavar = "genome_ref_path", dest = "rawGenome_path", required = True, help = "input path(s) to IMG reference genome files" )
 	parser.add_argument( "-o", metavar = "output_checked", dest = "checkedGenome_path", required = True, help = "path to output checked required genome files")
 	parser.add_argument( "-l", metavar = "log_file", dest = "log", required= True, help = "output log file" )
 	parser.add_argument( "-n", metavar = "min_length", dest = "imin", type = int, required = True, help = "minimal contig length" )
 
 	args = parser.parse_args()	
 	
-	rawGenome_path = args.rawGenome_path
+	arawGenome_path = map( lambda x: x+"/" if x[-1] != "/" else x, args.rawGenome_path.split(",") )
 	checkedGenome_path = args.checkedGenome_path
-	aastrLog = []
-
-	#This is needed for if you build a node by a path name
-	#there is no / at the end.
-	#Also add robustness against different user-defined
-	#reference genome path.
-	if rawGenome_path[-1] != "/":
-		rawGenome_path += "/"
 	if checkedGenome_path[-1] != "/":
 		checkedGenome_path += "/" 
+	
+	aastrLog = []
 	
 	#This is necessary for Scons will not create
 	#path if the node being built is a path.
@@ -86,10 +80,16 @@ def _main():
 		raise
 	
 	aastrLog.append( ["IMG Taxon ID", "# Total Contigs", "# Short Contigs"] )
-
 	for astrLine in csv.reader( refFile, csv.excel_tab ):
-		
-		strinput_Genome = rawGenome_path + astrLine[0]
+		strinput_Genome = ""
+		for rawGenome_path in arawGenome_path:
+			strinput_Genome = rawGenome_path + astrLine[0]
+			if os.path.isfile( strinput_Genome ):
+				break
+		if not strinput_Genome:
+			sys.stderr.write( "Cannot find genome reference file!\n" )
+			raise
+
 		stroutput_Genome = checkedGenome_path + astrLine[0]
 		if os.path.exists( stroutput_Genome ):
 			os.remove( stroutput_Genome )
